+++ app/plugins/pwa-install.client.ts
@@ -1,0 +1,5 @@
+import { bindPwaInstallListeners } from '~/composables/usePwaInstall'
+
+export default defineNuxtPlugin(() => {
+  bindPwaInstallListeners()
+})